backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg[binary]==3.2.5
python-dotenv==1.0.0
pydantic==2.5.2
pandas==2.2.3
scikit-learn==1.6.1
asyncpg
matplotlib

# cd API
#uvicorn main:app --reload --port 8084

#cd mspr
#npm run dev